+++ src/main/java/kr/hhplus/be/server/shared/aop/log/LogTrace.java
@@ -1,4 +1,4 @@
-package kr.hhplus.be.server.shared.aop;
+package kr.hhplus.be.server.shared.aop.log;
 
 public interface LogTrace {
