+++ src/main/java/kr/hhplus/be/server/shared/aop/log/ThreadLocalLogTrace.java
@@ -1,4 +1,4 @@
-package kr.hhplus.be.server.shared.aop;
+package kr.hhplus.be.server.shared.aop.log;
 
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.stereotype.Component;
