+++ src/main/java/kr/hhplus/be/server/shared/aop/log/TraceIdHolder.java
@@ -1,4 +1,4 @@
-package kr.hhplus.be.server.shared.aop;
+package kr.hhplus.be.server.shared.aop.log;
 
 public abstract class TraceIdHolder {
