+++ src/main/java/kr/hhplus/be/server/shared/aop/log/TraceStatus.java
@@ -1,4 +1,4 @@
-package kr.hhplus.be.server.shared.aop;
+package kr.hhplus.be.server.shared.aop.log;
 
 import lombok.AllArgsConstructor;
 import lombok.Getter;
